requirements.txt
celery==5.3.6
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
firebase-admin==6.5.0
graphene-django==3.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
pyfcm==1.5.4
pyJWT==2.8.0
python-decouple==3.8
redis==5.0.8
requests==2.31.0
